finish user routes & start bill and estimate route

+++ src/utils/globalUtils.ts
@@ -5,7 +5,7 @@ import { UserObject } from '../interfaces/userInterface';
  * Fonction de vérification des permissions pour une requête
  * @param user Utilisateur qui fait la requête
  * @param type Type requis pour faire la requête
- * @param role Rôle requis pour faire la requête
+ * @param admin Rôle requis pour faire la requête
  * @returns Retourne un booléen de si l'email existe ou non
  */
 const checkPermission = (user: UserObject, type: 'user' | 'client', admin?: boolean): boolean => {
@@ -27,7 +27,29 @@ const checkPermission = (user: UserObject, type: 'user' | 'client', admin?: bool
  * @param updateData Données à mettre à jour
  */
 const findOne = async (model: mongoose.Model<any, any>, id: string): Promise<any> => {
+    if (id.length !== 24) return null;
     return await model.findOne({ _id: mongoose.Types.ObjectId(id) });
+};
+
+/**
+ * Fonction pour trouver un document dans une collection et le populate.
+ * @param model Modèle mongoose
+ * @param id Id pour filtrer
+ * @param updateData Données à mettre à jour
+ */
+const findOneAndPopulate = async (model: mongoose.Model<any, any>, id: string, populate: string[]): Promise<any> => {
+    if (id.length !== 24) return null;
+    return await model.findOne({ _id: mongoose.Types.ObjectId(id) }).populate(populate);
+};
+
+/**
+ * Fonction pour trouver plusieurs documents dans une collection.
+ * @param model Modèle mongoose
+ * @param id Id pour filtrer
+ * @param updateData Données à mettre à jour
+ */
+const findMany = async (model: mongoose.Model<any, any>, filter: any): Promise<any> => {
+    return await model.find(filter);
 };
 
 /**
@@ -47,6 +69,7 @@ const updateOne = async (model: mongoose.Model<any, any>, filter: any, updateDat
  * @param updateData Données à mettre à jour
  */
 const updateOneById = async (model: mongoose.Model<any, any>, id: string, updateData: any): Promise<any> => {
+    if (id.length !== 24) return null;
     return await model.updateOne({ _id: mongoose.Types.ObjectId(id) }, { $set: updateData });
 };
 
@@ -56,11 +79,14 @@ const updateOneById = async (model: mongoose.Model<any, any>, id: string, update
  * @param id Id pour filtrer
  */
 const deleteOne = async (model: mongoose.Model<any, any>, id: string): Promise<any> => {
+    if (id.length !== 24) return null;
     return await model.deleteOne({ _id: mongoose.Types.ObjectId(id) });
 };
 
 const globalUtils = {
     findOne,
+    findOneAndPopulate,
+    findMany,
     updateOne,
     updateOneById,
     deleteOne,
